+++ src/main/kotlin/com/cognifide/gradle/aem/common/file/watcher/EventType.kt
@@ -1,0 +1,9 @@
+package com.cognifide.gradle.aem.common.file.watcher
+
+enum class EventType {
+    FILE_CREATED,
+    FILE_CHANGED,
+    FILE_DELETED,
+    DIR_CREATED,
+    DIR_DELETED,
+}
